Tighten compiler error/warning checks
Default --enable-warnings is now on.
  Use --disable-warnings if you can't fix your code.
Default for -Werror is now on.
  Use --disable-Werror if necessary.
  (Old old Bison needs it.)

+++ wafhelpers/check_sizeof.py
@@ -9,7 +9,7 @@ SIZE_FRAG = """
 %s
 #include <stdio.h>
 int main(void) {
-    printf("%%lu", sizeof(%s));
+    printf("%%lu", (unsigned long)sizeof(%s));
     return 0;
 }
 """
